drm: Handle race edgecase of MST connector removal
Handles the race edgecase where out_of_date gets set, but we are
currently inside of drm_prepare on the steamcompmgr thread.

Additionally, I am not even confident that this udev event for
device changed/removal is even indicative of the connector drm
object destruction.
It seems to not be related at all, I can always hit this path with MST
after waiting a bit after the first modeset then forcing another.
Which shouldn't ever happen if it was indicative of the connector
object destruction.